+++ aftiktuna-rust/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Idle animation for all creatures
 - New indoor decorations
+- Json location definition `crew_ship.json` for the ship and new symbol types "ship_controls" and "food_deposit"
 
 ### Changed